:recycle: refactor DefaultShoppingCartViewModel uiState 만들기
