+++ app/src/main/java/woowacourse/shopping/ui/productList/ProductListUiState.kt
@@ -4,6 +4,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import woowacourse.shopping.domain.model.Product
 
+abstract class ProductListUiState {
+    abstract val currentPage: LiveData<Int>
+    abstract val loadedProducts: LiveData<List<Product>>
+    abstract val productsHistory: LiveData<List<Product>>
+    abstract val cartProductTotalCount: LiveData<Int>
+    abstract val isLastPage: LiveData<Boolean>
+}
+
 data class DefaultProductListUiState(
     override val currentPage: MutableLiveData<Int> = MutableLiveData(FIRST_PAGE),
     override val loadedProducts: MutableLiveData<List<Product>> = MutableLiveData(emptyList()),
@@ -15,12 +23,4 @@ data class DefaultProductListUiState(
     companion object {
         private const val FIRST_PAGE = 1
     }
-}
-
-abstract class ProductListUiState {
-    abstract val currentPage: LiveData<Int>
-    abstract val loadedProducts: LiveData<List<Product>>
-    abstract val productsHistory: LiveData<List<Product>>
-    abstract val cartProductTotalCount: LiveData<Int>
-    abstract val isLastPage: LiveData<Boolean>
 }
