+++ app/src/main/res/layout/fragment_cart_list.xml
@@ -56,11 +56,11 @@
             android:id="@+id/cart_list_left_page_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:backgroundTint="@{vm.currentPage == 1 ? @color/gray : @color/black}"
-            android:enabled="@{vm.currentPage > 1}"
+            android:backgroundTint="@{vm.uiState.currentPage == 1 ? @color/gray : @color/black}"
+            android:enabled="@{vm.uiState.currentPage > 1}"
             android:onClick="@{() -> vm.previousPage()}"
             android:text="@string/previous_arrow"
-            android:textColor="@{vm.currentPage == 1 ? @color/gray : @color/white}"
+            android:textColor="@{vm.uiState.currentPage == 1 ? @color/gray : @color/white}"
             android:textSize="12sp"
             app:layout_constraintBottom_toBottomOf="@+id/cart_list_page_number"
             app:layout_constraintEnd_toStartOf="@id/cart_list_page_number"
@@ -75,7 +75,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="30dp"
             android:layout_marginBottom="50dp"
-            android:text="@{String.valueOf(vm.currentPage)}"
+            android:text="@{String.valueOf(vm.uiState.currentPage)}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toStartOf="@id/cart_list_right_page_button"
             app:layout_constraintStart_toEndOf="@id/cart_list_left_page_button"
@@ -87,11 +87,11 @@
             android:id="@+id/cart_list_right_page_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:backgroundTint="@{vm.isLastPage() ? @color/gray : @color/black}"
-            android:enabled="@{!vm.isLastPage()}"
+            android:backgroundTint="@{vm.uiState.isLastPage() ? @color/gray : @color/black}"
+            android:enabled="@{!vm.uiState.isLastPage()}"
             android:onClick="@{() -> vm.nextPage()}"
             android:text="@string/next_arrow"
-            android:textColor="@{vm.isLastPage() ? @color/gray : @color/white}"
+            android:textColor="@{vm.uiState.isLastPage() ? @color/gray : @color/white}"
             android:textSize="12sp"
             app:layout_constraintBottom_toBottomOf="@+id/cart_list_page_number"
             app:layout_constraintEnd_toEndOf="parent"
